refactor: Loading 컴포넌트 범용성 있게 개선
[#28]
